// src/app/modules/financeiro/planocontas/enums/financeiro-planocontas-enums.model.ts
/**********************************************************************************************
Code generated with MKL Plug-in version: 6.0.4
Code generated at time stamp: 2019-06-30T08:21:44.076
Copyright: Kerubin - logokoch@gmail.com

WARNING: DO NOT CHANGE THIS CODE BECAUSE THE CHANGES WILL BE LOST IN THE NEXT CODE GENERATION.
***********************************************************************************************/

import { SelectItem } from 'primeng/api';

export interface TipoReceitaDespesa extends SelectItem {
	
  label?: string;
  value: any;
  
}

export interface TipoPlanoContaFinanceiro extends SelectItem {
	
  label?: string;
  value: any;
  
}